[Commit - Bindings Package names]
Changing the names of the packages directory so to be consistent with name of the company .

It is to be noted that package name can not have hyphen "-" in the name.
So changing the directory names and putting under company names.

For example - dynamodb  package would be

dapr/components-contrib/bindings/aws/dynamodb