Fix body content parsing
This fixes a problem with the body of e-mail messages sent by the
SMTP binding.

When the SMTP binding was invoked over GRPC it would not translate the
data of the request to the right format in the e-mail. Instead it would
send an empty body. When the SMTP binding was invoked over HTTP it would
correctly unquote the data of the request.