dymamodb state: add support for first write concurrency
The existing logic for the dymamodb state store does not respect the
FirstWrite concurrency option for a SetRequest when etags are not used
by the client. This commit fixes the issue by employing a conditional
expression to test for the existence of an item when FirstWrite is
specified and etags are not.

Signed-off-by: Mike Brown <github@torvosoft.com>
Signed-off-by: Ryan Lettieri <ryanLettieri@microsoft.com>