Changes wasm middleware and output binding config to url instead of path
This switches the wasm middleware and output binding config to use url
instead of path to read the WebAssembly module. This allows us to later
implement file, http and OCI.

Note: The output binding passes STDIN as well as arguments. This means
it can run intepreters such as python. I will be implementing this with
the OCI scheme, tested with https://github.com/orgs/vmware-labs/packages/container/package/python-wasm

Signed-off-by: Adrian Cole <adrian@tetrate.io>