pubsub/rabbitmq: return error if reconnect errors in Init
We should always return the error if we fail to connection during the
Init of a Component. This is because failing to do so has downstream
effects whereby a component is considered healthy when it is not.

Re-attempting an Init or otherwise whould be handled by the runtime,
rather than on a per-component basis.

Signed-off-by: joshvanl <me@joshvanl.dev>