Improve Multi perf on SQL state stores with 1 op only
When Multi is invoked with a single operation, on SQL state stores (Postgres v1/v2, CockroachDB, MySQL, SQLite, MS SQL Server) we can omit starting a transaction, as a single operation is going to be executed atomically regardless. This allows reducing the number of round-trips to the DB from 3 to 1, reduces locks used inside the DB server, and improves overall performance.

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>