fix all wasm guests build and refresh with TinyGo 0.28.1
This fixes all the guests and align them to the same TinyGo
base version (0.28.1).

Specifically, we insert a go.mod for each guest project
to prevent mask out the `toolchain` directive in root go.mod

Finally, this fixes the Makefile for bindings/wasm for http;
the guest used is revlocked to TinyGo 0.28.1 for now.

Note: all the other guests build with latest TinyGo 0.31.1

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>