Working around current goavro bug (https://github.com/linkedin/goavro/issues/299 until merged).
I realized why this problem existed prior to 2.12 but never manifested in our code until 1.16.
This is because the old code was always creating a new codec when serializing (since it had to create a FullStandardJson codec). After the 1.16 change, it was reusing the schema.Codec() which would be subject to mutation issues...

For now, bringing back to old approach that always creates a new codec, which should get rid of the mutation issue altogether...

Signed-off-by: Patrick Assuied <patrick.assuied@elationhealth.com>